requirements.txt
Django==2.1.1
djangorestframework==3.8.2
gunicorn==19.9.0
jsonfield==2.0.2
pytz==2018.5
whitenoise==4.0
